Week_3/Azure_Deployment/requirements.txt
requests
streamlit
azure-ai-ml
azure.identity
